// src/interface/interface.ts
interface RouteInterface {
  path: string;
  component: any;
  isDefault?: boolean;
}